+++ src/main/java/com/example/voz_animal_proyect/usuario/service/UsuarioService.java
@@ -9,7 +9,7 @@ public interface UsuarioService {
 
      String findUsuarioByNombre(String nomUs, String contra);
 
-     Usuario obtenerUsuarioPorId(Integer id);
+     Usuario obtenerUsuarioPorId(long id);
 
      Usuario saveUser(Usuario usuario);
